Initial RobotBuilder commit
This is the initial commit of the Robot Project. All of the buttons on
the operator and driver joysticks are defined. I did not define the
buttons on the driver station because I don't know what positions they
are assigned at this time. Also, I have defined the chassis and elevator
subsystems, which may or may not be the only subsystems we need.

Also, since RobotBuilder doesn't cooperate 90% of the time, I created
templates so we don't have to even touch it. When we need to create a
new command or subsystem, we can just copy the template, rename it, and
go from there.