Added Debug methods
I added a debug mode thingy to Robot.java. I see us using this inside
commands to keep track of what's being run and to assist in debugging
autonomous programs and stuff. The debug mode just outputs a message
when it's enabled and does nothing when disabled. I also added a command
that we can run to turn debug mode on and off.