Merge for working code.
This branch was created so that we have an area for full test code, even
though we are not ready to merge to master. This is currently a merge of
the intake, lift, and drive branches. This branch was created as a copy
of drive, so it already had drive code. We then added each command and
subsystem not included, and edited OI and RobotMap to include the added
subsystems.