Fixed motor inversion
I could've sworn I already fixed this, but I guess not... Here I added
motor inversion commands to each of the drive methods so each of the
different methods drives correctly. There's definitely a more efficient
way to do this than setting motor inversion every time we update the
motors, but hey, it works.

Also, as far as I can tell, there are no problems with the MASTER
branch at this point in time. I've been testing for a while and I
haven't found any errors so far.