Added Drive, Strafe, and Wait commands
Here I added 2 driving commands and a wait command. The 2 driving
commands, AutonDriveAtSpeedForTime and AutonStrafeAtSpeedForTime, take a
move/strafe variable ranging from -1.0 (backward) to 1.0 (forward) and a
time variable in seconds. The AutonWait command just takes a time
variable, in seconds, and waits for the specified time before allowing
the next scheduled command to run.