Added tote sensor commands & tweaked Mecanum inversion
I added 2 commands: one to drive until the tote sensor is triggered and
another one to drive until the tote sensor is not triggered. I also
tweaked the mecanum value inversion to make more sense (a.k.a. It
doesn't autocorrect for joystick inversion anymore, but I moved that to
the joystick command itself.)