Defined the Gyro w/ methods & robotPeriodic
Here I defined the Gyro in RobotMap and I have it getting zeroed
automatically as soon as calibration is complete.  Then I created a few
methods in chassis.java to return if the gyro is currently calibrating,
reset the gyro, and read the gyro's current angle (gyroIsCalibrating,
gyroReset, and gyroReadAngle respectively). The gyroReset method has
also been extended into a standalone command called GyroReset.java.

I also created a robotPeriodic method in Robot.java that's called in the
periodic method for every mode. This provides us an overall periodic
method we can use. Right now this contains SmartDashboard updates for
the gyroscope.