Added AutonDriveTurnDegrees command & test
I created an AutonDriveTurnDegrees command (long name... I know...) that
turns the robot an approximate number of degrees at a specified speed. I
also added an excessive amount of test commands for this to the
AutonomousCommand command. The turning command seems to be working, but
I can't be positive yet because the robot has no wheels and I can't tell
which way the motors are rotating. So yea.