Updated LED code for one less strip.
Since we have one less tote, we are consolidating the finger light
sensor and the kicker out sensor into one. We only need to know whether
the kicker is out when we are in override, which is when we don't need
to see the finger sensor. If we are in override, the finger sensor LEDs
will become set to the kicker out LED, and will flash when the kicker is
out.