Merged in kicker and LED
We merged in all of the LED code from the LED branch, and all of the
kicker bug fixes/updates from both the kicker and LED branches (some
commits were misplaced during testing, and bunched into the LED branch).
LEDs and kicker should now function, as long as they are set to the
proper ports in this branch (I'm not quite sure whether this branch has
the proper port numbers or not. I changed some, but chances are more are
still done wrong).