Code cleanliness and organization, new pneumatics subsystems.
We moved all of our intake pneumatics from the intake subsystem into
their own new IntakePneumatics subsystem. We also added our compressor
commands, HOWEVER, the start command still needs to be attached to a
button, we merely don't know which one yet. Also, there is a lot of
cleaning in this commit, due to misspellings and other minor problems.