Auton and Kicker
Modified Auton J to work with the strafing encoder
Updated Kicker commands for in and out to be accurate.
Removed auto tote sensor and finger sensor.
Cleaned up a bunch of left over code in Robot. Robot Map, and  OI
Removed the LED entries for auto tote sensor and finger sensor.
Deleted a bunch of auto commands.