Comments and update of button locations
Entered Javadocs in the subsystems for constructors.
Created CONTAINERARM_JoystickThrottle for using the throttle on the
Logitech.
Moved the bottom position on LIFT_JoystickControl to adapt to the
handle position for the tote.
Modified names and commented buttons in OI.
Cleaned up imports In Robot.
Changed the default position in ContainerArmLift.